fixed transit queries not using filesystemtype \n removed default file system type

+++ lib/Youverse.Core.Storage/SQLite/DriveDatabase/TableCommandMessageQueue.cs
@@ -58,7 +58,7 @@ namespace Youverse.Core.Storage.Sqlite.DriveDatabase
                                 throw new Exception("Not a guid");
                             ts = rdr.GetInt64(1);
 
-                            queue.Add(new CommandMessage { fileId = new Guid(_guid), timestamp = new UnixTimeUtc((UInt64)ts) });
+                            queue.Add(new CommandMessage { fileId = new Guid(_guid), timestamp = new UnixTimeUtc(ts) });
                             i++;
                         }
